(refactor) Remove converting-model package completely
