+++ code/features-convert/pubdate/src/main/java/nu/marginalia/pubdate/PubDateHeuristic.java
@@ -2,7 +2,7 @@ package nu.marginalia.pubdate;
 
 import nu.marginalia.model.EdgeUrl;
 import nu.marginalia.model.crawl.PubDate;
-import nu.marginalia.converting.model.HtmlStandard;
+import nu.marginalia.model.html.HtmlStandard;
 import org.jsoup.nodes.Document;
 
 import java.util.Optional;
