--- code/process-models/converting-model/build.gradle
@@ -1,47 +1,0 @@
-plugins {
-    id 'java'
-    id "io.freefair.lombok" version "8.2.2"
-
-    id 'jvm-test-suite'
-}
-
-java {
-    toolchain {
-        languageVersion.set(JavaLanguageVersion.of(20))
-    }
-}
-dependencies {
-
-    //implementation project(':third-party:monkey-patch-gson')
-
-    implementation project(':code:common:db')
-    implementation project(':code:common:model')
-    implementation project(':code:api:index-api')
-    implementation project(':code:common:service-discovery')
-    implementation project(':code:common:service-client')
-    implementation project(':code:libraries:language-processing')
-
-    implementation project(':code:features-convert:keyword-extraction')
-
-    implementation libs.lombok
-    annotationProcessor libs.lombok
-    implementation libs.bundles.slf4j
-
-    implementation libs.notnull
-    implementation libs.trove
-    implementation libs.fastutil
-
-    testImplementation libs.bundles.slf4j.test
-    testImplementation libs.bundles.junit
-    testImplementation libs.mockito
-}
-
-test {
-    useJUnitPlatform()
-}
-
-task fastTests(type: Test) {
-    useJUnitPlatform {
-        excludeTags "slow"
-    }
-}
