--- code/process-models/converting-model/readme.md
@@ -1,3 +1,0 @@
-# Converting Models
-
-!!To be deleted!!
