+++ code/processes/converting-process/build.gradle
@@ -43,7 +43,6 @@ dependencies {
     implementation project(':code:libraries:big-string')
     implementation project(':code:libraries:language-processing')
 
-    implementation project(':code:process-models:converting-model')
     implementation project(':code:process-models:processed-data')
     implementation project(':code:process-models:work-log')
     implementation project(':code:process-models:crawling-model')
