+++ code/processes/loading-process/build.gradle
@@ -35,7 +35,6 @@ dependencies {
     testImplementation project(':code:services-core:search-service')
 
     implementation project(':code:process-models:crawling-model')
-    implementation project(':code:process-models:converting-model')
     implementation project(':code:process-models:processed-data')
     implementation project(':code:process-models:work-log')
